Moved landing page components to new folder
