+++ components/landing/landing-header.js
@@ -1,7 +1,7 @@
 import LandingNavbarLinks from './landing-navbar-links';
 import Link from 'next/link';
-import MenuIcon from './menu-icon';
-import utilStyles from '../styles/utils.module.css';
+import MenuIcon from '../menu-icon';
+import utilStyles from '../../styles/utils.module.css';
 import landingHeaderStyles from './landing-header.module.css';
 
 const name = 'Rushing Labs';
