+++ components/landing/landing-layout.js
@@ -1,12 +1,8 @@
 import LandingNavbarLinks from './landing-navbar-links';
-import Link from 'next/link';
-import MenuIcon from './menu-icon';
-import LayoutHeadLoader from './layout-head-loader';
+import LayoutHeadLoader from '../layout-head-loader';
 import LandingHeader from './landing-header';
 
 import landingLayoutStyles from './landing-layout.module.css';
-import styles from './layout.module.css';
-import utilStyles from '../styles/utils.module.css';
 
 const name = 'Rushing Labs';
