+++ components/landing/landing-navbar-links.js
@@ -1,5 +1,5 @@
-import styles from './layout.module.css';
-import utilStyles from '../styles/utils.module.css';
+import styles from '../layout.module.css';
+import utilStyles from '../../styles/utils.module.css';
 import Link from 'next/link';
 
 export default function LandingNavbarLinks({}) {
